# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shared memory layer for security operations: episodic observations, versioned facts, hybrid recall, graph traversal.",
        "title": "SOC Memory Layer API",
        "contact": {
            "name": "SOC Platform Team"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/enrich": {
            "get": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Entity card for one key: profile, facts, observations, and graph neighbors.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recall"
                ],
                "summary": "Enrich an entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity type (e.g. ioc_ip, ioc_domain)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Entity lookup key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity card; found=false with empty lists when unknown",
                        "schema": {
                            "$ref": "#/definitions/internal_api.enrichResp"
                        }
                    },
                    "400": {
                        "description": "Missing type/key or storage validation failure",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/facts": {
            "post": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Assert a subject-predicate-object fact; agents propose, humans write active versions directly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write"
                ],
                "summary": "Assert a fact",
                "parameters": [
                    {
                        "description": "Fact payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.assertFactReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Written fact (proposed or active per actor type)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.factResp"
                        }
                    },
                    "400": {
                        "description": "Invalid body or oversized predicate/value",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "429": {
                        "description": "Agent write budget exhausted (Retry-After set)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/facts/{id}/promote": {
            "post": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Transition a proposed fact to active; human-gated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write"
                ],
                "summary": "Promote a fact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fact ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promoted fact with the current version ID",
                        "schema": {
                            "$ref": "#/definitions/internal_api.transitionResp"
                        }
                    },
                    "400": {
                        "description": "Invalid fact ID",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "403": {
                        "description": "Only human actors may promote (human_gated)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "404": {
                        "description": "Fact not found or ID consumed by an earlier transition",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "409": {
                        "description": "Fact is not in proposed status",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "429": {
                        "description": "Agent write budget exhausted (Retry-After set)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/facts/{id}/retract": {
            "post": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Mark a fact retracted with an optional reason; human-gated. Empty request body allowed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write"
                ],
                "summary": "Retract a fact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fact ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional reason",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.retractReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retracted fact with the current version ID",
                        "schema": {
                            "$ref": "#/definitions/internal_api.transitionResp"
                        }
                    },
                    "400": {
                        "description": "Invalid fact ID or malformed body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "403": {
                        "description": "Only human actors may retract (human_gated)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "404": {
                        "description": "Fact not found or ID consumed by an earlier transition",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "409": {
                        "description": "Fact already retracted",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "429": {
                        "description": "Agent write budget exhausted (Retry-After set)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api.assertFactReq": {
            "type": "object",
            "properties": {
                "client_event_id": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "object_id": {
                    "type": "string"
                },
                "object_value": {
                    "type": "string"
                },
                "predicate": {
                    "type": "string"
                },
                "source_obs": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.enrichResp": {
            "type": "object",
            "properties": {
                "entity": {
                    "$ref": "#/definitions/internal_api.entityJSON"
                },
                "facts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.factViewJSON"
                    }
                },
                "found": {
                    "type": "boolean"
                },
                "neighbors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.neighborJSON"
                    }
                },
                "observations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.obsViewJSON"
                    }
                }
            }
        },
...
        "internal_api.errBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_api.errInner"
                }
            }
        },
...
        "internal_api.factResp": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "object_value": {
                    "type": "string"
                },
                "predicate": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "written_by": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.transitionResp": {
            "type": "object",
            "properties": {
                "fact": {
                    "$ref": "#/definitions/internal_api.factResp"
                },
                "transitioned_by": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "IdentityHeaders": {
            "description": "Plus X-Actor-ID and X-Scope headers; identity never travels in bodies.",
            "type": "apiKey",
            "name": "X-Actor-Type",
            "in": "header"
        }
    }
}